Move type registry into documents, fix circular import
